requirements.txt
# Core dependencies
adblockparser==0.7
beautifulsoup4==4.13.4
cryptography>=43.0.5,<46.0.0
dnspython>=2.6.1,<3.0.0
elasticsearch==9.0.2
exifread==3.3.1
ipwhois>=1.3.0,<2.0.0
lxml>=5.4.0,<7.0.0
netaddr>=1.3.0,<2.0.0
networkx>=3.2.1,<4.0.0
openpyxl>=3.1.5,<4.0.0
paramiko>=3.0.0,<4.0.0
phonenumbers>=9.0.7,<10.0.0
psycopg2-binary>=2.9.3,<3.0.0
publicsuffixlist>=1.0.2.20250613,<2.0.0
pygexf>=0.2.2,<0.3.0
pyopenssl>=24.0.0,<26.0.0
pypdf>=3.1.0,<6.0.0
pysocks>=1.7.1,<2.0.0
python-docx>=1.1.2,<2.0.0
python-pptx>=1.0.2,<2.0.0
python-whois>=0.9.5,<1.0.0
pyyaml>=6.0.1,<7.0.0
requests>=2.32.0,<3.0.0
secure>=0.3.0,<1.0.0
jsonschema>=4.0.0,<5.0.0
pillow>=10.3.0,<12.0.0
setuptools>=78.1.1,<81.0.0
zipp>=3.19.1,<4.0.0

# FastAPI & web stack
fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.18,<1.0.0
websockets>=12.0,<14.0.0
httpx>=0.24.0,<1.0.0
markdown>=3.4.0,<4.0.0
aiohttp>=3.9.0,<4.0.0

# Auth & security
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
pyjwt>=2.8.0,<3.0.0
bcrypt>=4.0.0,<5.0.0
ldap3>=2.9.0,<3.0.0
werkzeug>=2.3.0,<3.0.0
slowapi>=0.1.9,<1.0.0
email-validator>=2.0.0,<3.0.0
bleach>=6.0.0,<7.0.0

# Data & integrations
redis>=5.0.0,<6.0.0
openai>=1.0.0,<2.0.0
telethon>=1.36.0,<2.0.0
python-dateutil>=2.8.0,<3.0.0
minio>=7.2.0,<8.0.0

# GraphQL
strawberry-graphql[fastapi]>=0.235.0,<1.0.0

# gRPC inter-service communication
grpcio>=1.60.0,<2.0.0
grpcio-tools>=1.60.0,<2.0.0
protobuf>=4.25.0,<6.0.0

# Celery task queue + scheduling
celery[redis]>=5.4.0,<6.0.0
flower>=2.0.0,<3.0.0

# PDF report generation
weasyprint>=62.0,<64.0.0
jinja2>=3.1.0,<4.0.0
